+++ CHANGELOG.md
@@ -6,6 +6,16 @@
 
 * Fix: [#69](https://github.com/savonrb/httpi/issues/69) truncating response headers.
 
+## 2.0.2 (2013-01-26)
+
+* Feature: Changed `HTTPI::Request#set_cookies` to accept an Array of `HTTPI::Cookie`
+  objects as well as any object that responds to `cookies` (like an `HTTPI::Response`).
+
+## 2.0.1 (2013-01-25)
+
+* Fix: [#72](https://github.com/savonrb/httpi/pull/72) standardized response
+  headers from all adapters.
+
 ## 2.0.0 (2012-12-16)
 
 * Feature: [#66](https://github.com/savonrb/httpi/pull/66) adds a `query` method